Rework package and file structure
